Refactor processing of error messages to AntlrDocument
