+++ src/org/github/antlr4ide/editor/ANTLRv4DocumentProvider.java
@@ -11,7 +11,6 @@ import org.eclipse.jface.text.rules.FastPartitioner;
 import org.eclipse.ui.IEditorInput;
 import org.eclipse.ui.IFileEditorInput;
 import org.eclipse.ui.editors.text.FileDocumentProvider;
-import org.eclipse.ui.editors.text.TextEditor;
 
 public class ANTLRv4DocumentProvider extends FileDocumentProvider {
 
@@ -20,7 +19,7 @@ public class ANTLRv4DocumentProvider extends FileDocumentProvider {
 	@Deprecated
 	private ANTLRv4Editor editor;
 	
-	private IDocument doc;
+	private AntlrDocument doc;
 
 
 	public ANTLRv4DocumentProvider(ANTLRv4Editor antlRv4Editor) { 
@@ -30,6 +29,7 @@ public class ANTLRv4DocumentProvider extends FileDocumentProvider {
 	}
 
 
+	@Override
 	protected IDocument createDocument(Object element) throws CoreException {
 		if (DEBUG) {
 			System.out.println(">>> ANTLRv4DocumentProvider.createDocument element>" + element.getClass() + "<");
@@ -44,7 +44,8 @@ public class ANTLRv4DocumentProvider extends FileDocumentProvider {
 		}
 		
 		IDocument document = super.createDocument(element);
-		document.addDocumentListener(new MyDocumentListener(editor));
+		document.addDocumentListener(new AntlrDocument.AntlrDocumentChangedListener(editor));
+		this.doc=(AntlrDocument) document;
 		
 		if (document != null) {
 			IDocumentPartitioner partitioner =
@@ -55,13 +56,18 @@ public class ANTLRv4DocumentProvider extends FileDocumentProvider {
 						});
 			partitioner.connect(document);
 			document.setDocumentPartitioner(partitioner);
-			((AntlrDocument)document).scan(); // set the initial list of tokens.
+			doc.setEditor(this.editor);
+			if(document.getLength()>0)
+			  this.doc.scan(); // set the initial list of tokens.
 		}
-		this.doc=document;
 		return document;
 	}
 	
-	
+
+	/**
+	 * Ensure the IDocument is based on AntlrDocument
+	 */
+	@Override
 	protected IDocument createEmptyDocument () {
 		return new AntlrDocument();
 	}
@@ -72,50 +78,4 @@ public class ANTLRv4DocumentProvider extends FileDocumentProvider {
 	}
 
 
-	// TODO: Move to AntlrScanner
-	public class MyDocumentListener implements IDocumentListener  {
-		TextEditor editor;
-
-		public MyDocumentListener(ANTLRv4Editor antlRv4Editor) {
-			this.editor=antlRv4Editor;
-		}
-
-		@Override
-		public void documentAboutToBeChanged(DocumentEvent event) {
-			// nothing
-		}
-
-		@Override
-		public void documentChanged(DocumentEvent event) {
-//			System.out.println("documentChanged " +event);
-			try {
-			AntlrDocument doc=(AntlrDocument)  event.getDocument();
-			doc.scan();
-
-			// set any error markers from the scanner
-			IEditorInput input = this.editor.getEditorInput();
-			IFile file = ((IFileEditorInput) input).getFile();
-			file.deleteMarkers(IMarker.PROBLEM, false, 1);
-			for (String err : doc.errorList) {
-				// annotate resource with errors.
-				// errorlist format line:position:token.getStartIndex():token.getStopIndex():message
-				String ss[] = err.split(":");
-				int line = Integer.parseInt(ss[0]);
-				int start = Integer.parseInt(ss[2]);
-				int end = Integer.parseInt(ss[3]) + 1; // token.getStopIndex() is inclusive. char_end is exclusive so add +1
-
-				String mrkAttr[] = { IMarker.SEVERITY, IMarker.LINE_NUMBER, IMarker.MESSAGE, IMarker.CHAR_START, IMarker.CHAR_END };
-				Object mrkValue[] = { IMarker.SEVERITY_ERROR, line, ss[4], start, end };
-				IMarker marker = file.createMarker(IMarker.PROBLEM);
-				marker.setAttributes(mrkAttr, mrkValue);
-			}
-		}
-		 catch (CoreException e) {
-			e.printStackTrace();
-		}
-		
-	}
-
-	}
-	
 }
